Внесены правки в структуру приложения. Выполнена подготовка к релизу.
