+++ TestTask/Helpers/CharTypeHelper.cs
@@ -1,11 +1,8 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace TestTask.Helpers
+﻿namespace TestTask.Helpers
 {
+    /// <summary>
+    /// Представляет набор констант для анализа букв на предмет гласности, согласности.
+    /// </summary>
     internal static class CharTypeHelper
     {
         /// <summary>
@@ -17,15 +14,5 @@ namespace TestTask.Helpers
         /// Набор согласных букв для русского языка.
         /// </summary>
         internal const string RUS_EN_CONSONANTS = "БВГДЖЗЙКЛМНПРСТФХЦЧШЩBCDFGHJKLMNPQRSTVWXZ";
-
-        /// <summary>
-        /// Набор гласных букв для английского языка.
-        /// </summary>
-        internal const string EN_VOWELS = "AEIOUY";
-
-        /// <summary>
-        /// Набор согласных букв для английского языка.
-        /// </summary>
-        internal const string EN_CONSONANTS = "BCDFGHJKLMNPQRSTVWXZ";
     }
 }
